use gpio 0 instead of 16
gpio16 is used for waking up from deep sleep;
it is better to not use this for soft serial.

gpio 0 doesn't have any special functions so we
can use this instead.

gpio2 is connected to the led and it's always
useful to have a visual indication of whether
there is data that is being received